- #61: Moving WarTier and Tech files under techs/ folder.
